// graphql/server/src/codegen/orm/input-types.ts
export interface StringFilter {
  isNull?: boolean;
  equalTo?: string;
  notEqualTo?: string;
  distinctFrom?: string;
  notDistinctFrom?: string;
  in?: string[];
  notIn?: string[];
  lessThan?: string;
  lessThanOrEqualTo?: string;
  greaterThan?: string;
  greaterThanOrEqualTo?: string;
  includes?: string;
  notIncludes?: string;
  includesInsensitive?: string;
  notIncludesInsensitive?: string;
  startsWith?: string;
  notStartsWith?: string;
  startsWithInsensitive?: string;
  notStartsWithInsensitive?: string;
  endsWith?: string;
  notEndsWith?: string;
  endsWithInsensitive?: string;
  notEndsWithInsensitive?: string;
  like?: string;
  notLike?: string;
  likeInsensitive?: string;
  notLikeInsensitive?: string;
}
export interface IntFilter {
  isNull?: boolean;
  equalTo?: number;
  notEqualTo?: number;
  distinctFrom?: number;
  notDistinctFrom?: number;
  in?: number[];
  notIn?: number[];
  lessThan?: number;
  lessThanOrEqualTo?: number;
  greaterThan?: number;
  greaterThanOrEqualTo?: number;
}
export interface FloatFilter {
  isNull?: boolean;
  equalTo?: number;
  notEqualTo?: number;
  distinctFrom?: number;
  notDistinctFrom?: number;
  in?: number[];
  notIn?: number[];
  lessThan?: number;
  lessThanOrEqualTo?: number;
  greaterThan?: number;
  greaterThanOrEqualTo?: number;
}
export interface BooleanFilter {
  isNull?: boolean;
  equalTo?: boolean;
  notEqualTo?: boolean;
}
export interface UUIDFilter {
  isNull?: boolean;
  equalTo?: string;
  notEqualTo?: string;
  distinctFrom?: string;
  notDistinctFrom?: string;
  in?: string[];
  notIn?: string[];
}
export interface DatetimeFilter {
  isNull?: boolean;
  equalTo?: string;
  notEqualTo?: string;
  distinctFrom?: string;
  notDistinctFrom?: string;
  in?: string[];
  notIn?: string[];
  lessThan?: string;
  lessThanOrEqualTo?: string;
  greaterThan?: string;
  greaterThanOrEqualTo?: string;
}
export interface DateFilter {
  isNull?: boolean;
  equalTo?: string;
  notEqualTo?: string;
  distinctFrom?: string;
  notDistinctFrom?: string;
  in?: string[];
  notIn?: string[];
  lessThan?: string;
  lessThanOrEqualTo?: string;
  greaterThan?: string;
  greaterThanOrEqualTo?: string;
}
export interface JSONFilter {
  isNull?: boolean;
  equalTo?: Record<string, unknown>;
  notEqualTo?: Record<string, unknown>;
  distinctFrom?: Record<string, unknown>;
  notDistinctFrom?: Record<string, unknown>;
  contains?: Record<string, unknown>;
  containedBy?: Record<string, unknown>;
  containsKey?: string;
  containsAllKeys?: string[];
  containsAnyKeys?: string[];
}
// ...
export interface Database {
  id: string;
  ownerId?: string | null;
  schemaName?: string | null;
  privateSchemaName?: string | null;
  name?: string | null;
  label?: string | null;
  hash?: string | null;
  createdAt?: string | null;
  updatedAt?: string | null;
}
// ...
export interface ApiExtension {
  id: string;
  schemaName?: string | null;
  databaseId?: string | null;
  apiId?: string | null;
}
export interface ApiModule {
  id: string;
  databaseId?: string | null;
  apiId?: string | null;
  name?: string | null;
  data?: Record<string, unknown> | null;
}
export interface Domain {
  id: string;
  databaseId?: string | null;
  apiId?: string | null;
  siteId?: string | null;
  subdomain?: string | null;
  domain?: string | null;
}
// ...
export interface RlsModule {
  id: string;
  databaseId?: string | null;
  apiId?: string | null;
  schemaId?: string | null;
  privateSchemaId?: string | null;
  tokensTableId?: string | null;
  usersTableId?: string | null;
  authenticate?: string | null;
  authenticateStrict?: string | null;
  currentRole?: string | null;
  currentRoleId?: string | null;
}
// ...
export interface Api {
  id: string;
  databaseId?: string | null;
  name?: string | null;
  dbname?: string | null;
  roleName?: string | null;
  anonRole?: string | null;
  isPublic?: boolean | null;
}
// ...
export interface ConnectionResult<T> {
  nodes: T[];
  totalCount: number;
  pageInfo: PageInfo;
}
// ...
export interface DomainRelations {
  database?: Database | null;
  api?: Api | null;
  site?: Site | null;
}
// ...
export interface ApiRelations {
  database?: Database | null;
  rlsModule?: RlsModule | null;
  apiExtensions?: ConnectionResult<ApiExtension>;
  apiModules?: ConnectionResult<ApiModule>;
  apiSchemas?: ConnectionResult<ApiSchema>;
  domains?: ConnectionResult<Domain>;
  databasesByApiExtensionApiIdAndDatabaseId?: ConnectionResult<Database>;
  schemasByApiSchemaApiIdAndSchemaId?: ConnectionResult<Schema>;
}
// ...
export type DomainWithRelations = Domain & DomainRelations;
// ...
export type ApiWithRelations = Api & ApiRelations;
// ...
export type ApiSchemaSelect = {
  id?: boolean;
  databaseId?: boolean;
  schemaId?: boolean;
  apiId?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  schema?:
    | boolean
    | {
        select?: SchemaSelect;
      };
  api?:
    | boolean
    | {
        select?: ApiSelect;
      };
};
export type ApiExtensionSelect = {
  id?: boolean;
  schemaName?: boolean;
  databaseId?: boolean;
  apiId?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  api?:
    | boolean
    | {
        select?: ApiSelect;
      };
};
export type ApiModuleSelect = {
  id?: boolean;
  databaseId?: boolean;
  apiId?: boolean;
  name?: boolean;
  data?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  api?:
    | boolean
    | {
        select?: ApiSelect;
      };
};
export type DomainSelect = {
  id?: boolean;
  databaseId?: boolean;
  apiId?: boolean;
  siteId?: boolean;
  subdomain?: boolean;
  domain?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  api?:
    | boolean
    | {
        select?: ApiSelect;
      };
  site?:
    | boolean
    | {
        select?: SiteSelect;
      };
};
// ...
export type RlsModuleSelect = {
  id?: boolean;
  databaseId?: boolean;
  apiId?: boolean;
  schemaId?: boolean;
  privateSchemaId?: boolean;
  tokensTableId?: boolean;
  usersTableId?: boolean;
  authenticate?: boolean;
  authenticateStrict?: boolean;
  currentRole?: boolean;
  currentRoleId?: boolean;
  api?:
    | boolean
    | {
        select?: ApiSelect;
      };
  schema?:
    | boolean
    | {
        select?: SchemaSelect;
      };
  privateSchema?:
    | boolean
    | {
        select?: SchemaSelect;
      };
};
// ...
export type ApiSelect = {
  id?: boolean;
  databaseId?: boolean;
  name?: boolean;
  dbname?: boolean;
  roleName?: boolean;
  anonRole?: boolean;
  isPublic?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  rlsModule?:
    | boolean
    | {
        select?: RlsModuleSelect;
      };
  apiExtensions?:
    | boolean
    | {
        select?: ApiExtensionSelect;
        first?: number;
        filter?: ApiExtensionFilter;
        orderBy?: ApiExtensionsOrderBy[];
      };
  apiModules?:
    | boolean
    | {
        select?: ApiModuleSelect;
        first?: number;
        filter?: ApiModuleFilter;
        orderBy?: ApiModulesOrderBy[];
      };
  apiSchemas?:
    | boolean
    | {
        select?: ApiSchemaSelect;
        first?: number;
        filter?: ApiSchemaFilter;
        orderBy?: ApiSchemasOrderBy[];
      };
  domains?:
    | boolean
    | {
        select?: DomainSelect;
        first?: number;
        filter?: DomainFilter;
        orderBy?: DomainsOrderBy[];
      };
  databasesByApiExtensionApiIdAndDatabaseId?:
    | boolean
    | {
        select?: DatabaseSelect;
        first?: number;
        filter?: DatabaseFilter;
        orderBy?: DatabasesOrderBy[];
      };
  schemasByApiSchemaApiIdAndSchemaId?:
    | boolean
    | {
        select?: SchemaSelect;
        first?: number;
        filter?: SchemaFilter;
        orderBy?: SchemasOrderBy[];
      };
};
export type SiteSelect = {
  id?: boolean;
  databaseId?: boolean;
  title?: boolean;
  description?: boolean;
  ogImage?: boolean;
  favicon?: boolean;
  appleTouchIcon?: boolean;
  logo?: boolean;
  dbname?: boolean;
  database?:
    | boolean
    | {
        select?: DatabaseSelect;
      };
  app?:
    | boolean
    | {
        select?: AppSelect;
      };
  domains?:
    | boolean
    | {
        select?: DomainSelect;
        first?: number;
        filter?: DomainFilter;
        orderBy?: DomainsOrderBy[];
      };
  siteMetadata?:
    | boolean
    | {
        select?: SiteMetadatumSelect;
        first?: number;
        filter?: SiteMetadatumFilter;
        orderBy?: SiteMetadataOrderBy[];
      };
  siteModules?:
    | boolean
    | {
        select?: SiteModuleSelect;
        first?: number;
        filter?: SiteModuleFilter;
        orderBy?: SiteModulesOrderBy[];
      };
  siteThemes?:
    | boolean
    | {
        select?: SiteThemeSelect;
        first?: number;
        filter?: SiteThemeFilter;
        orderBy?: SiteThemesOrderBy[];
      };
};
// ...
export interface DatabaseFilter {
  id?: UUIDFilter;
  ownerId?: UUIDFilter;
  schemaName?: StringFilter;
  privateSchemaName?: StringFilter;
  name?: StringFilter;
  label?: StringFilter;
  hash?: UUIDFilter;
  createdAt?: DatetimeFilter;
  updatedAt?: DatetimeFilter;
  and?: DatabaseFilter[];
  or?: DatabaseFilter[];
  not?: DatabaseFilter;
}
// ...
export interface SchemaFilter {
  id?: UUIDFilter;
  databaseId?: UUIDFilter;
  name?: StringFilter;
  schemaName?: StringFilter;
  label?: StringFilter;
  description?: StringFilter;
  smartTags?: JSONFilter;
  category?: StringFilter;
  module?: StringFilter;
  scope?: IntFilter;
  tags?: StringFilter;
  isPublic?: BooleanFilter;
  createdAt?: DatetimeFilter;
  updatedAt?: DatetimeFilter;
  and?: SchemaFilter[];
  or?: SchemaFilter[];
  not?: SchemaFilter;
}
// ...
export interface ApiSchemaFilter {
  id?: UUIDFilter;
  databaseId?: UUIDFilter;
  schemaId?: UUIDFilter;
  apiId?: UUIDFilter;
  and?: ApiSchemaFilter[];
  or?: ApiSchemaFilter[];
  not?: ApiSchemaFilter;
}
export interface ApiExtensionFilter {
  id?: UUIDFilter;
  schemaName?: StringFilter;
  databaseId?: UUIDFilter;
  apiId?: UUIDFilter;
  and?: ApiExtensionFilter[];
  or?: ApiExtensionFilter[];
  not?: ApiExtensionFilter;
}
export interface ApiModuleFilter {
  id?: UUIDFilter;
  databaseId?: UUIDFilter;
  apiId?: UUIDFilter;
  name?: StringFilter;
  data?: JSONFilter;
  and?: ApiModuleFilter[];
  or?: ApiModuleFilter[];
  not?: ApiModuleFilter;
}
export interface DomainFilter {
  id?: UUIDFilter;
  databaseId?: UUIDFilter;
  apiId?: UUIDFilter;
  siteId?: UUIDFilter;
  subdomain?: StringFilter;
  domain?: StringFilter;
  and?: DomainFilter[];
  or?: DomainFilter[];
  not?: DomainFilter;
}
// ...
export type DatabasesOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'OWNER_ID_ASC'
  | 'OWNER_ID_DESC'
  | 'SCHEMA_NAME_ASC'
  | 'SCHEMA_NAME_DESC'
  | 'PRIVATE_SCHEMA_NAME_ASC'
  | 'PRIVATE_SCHEMA_NAME_DESC'
  | 'NAME_ASC'
  | 'NAME_DESC'
  | 'LABEL_ASC'
  | 'LABEL_DESC'
  | 'HASH_ASC'
  | 'HASH_DESC'
  | 'CREATED_AT_ASC'
  | 'CREATED_AT_DESC'
  | 'UPDATED_AT_ASC'
  | 'UPDATED_AT_DESC';
// ...
export type SchemasOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'DATABASE_ID_ASC'
  | 'DATABASE_ID_DESC'
  | 'NAME_ASC'
  | 'NAME_DESC'
  | 'SCHEMA_NAME_ASC'
  | 'SCHEMA_NAME_DESC'
  | 'LABEL_ASC'
  | 'LABEL_DESC'
  | 'DESCRIPTION_ASC'
  | 'DESCRIPTION_DESC'
  | 'SMART_TAGS_ASC'
  | 'SMART_TAGS_DESC'
  | 'CATEGORY_ASC'
  | 'CATEGORY_DESC'
  | 'MODULE_ASC'
  | 'MODULE_DESC'
  | 'SCOPE_ASC'
  | 'SCOPE_DESC'
  | 'TAGS_ASC'
  | 'TAGS_DESC'
  | 'IS_PUBLIC_ASC'
  | 'IS_PUBLIC_DESC'
  | 'CREATED_AT_ASC'
  | 'CREATED_AT_DESC'
  | 'UPDATED_AT_ASC'
  | 'UPDATED_AT_DESC';
// ...
export type ApiSchemasOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'DATABASE_ID_ASC'
  | 'DATABASE_ID_DESC'
  | 'SCHEMA_ID_ASC'
  | 'SCHEMA_ID_DESC'
  | 'API_ID_ASC'
  | 'API_ID_DESC';
export type ApiExtensionsOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'SCHEMA_NAME_ASC'
  | 'SCHEMA_NAME_DESC'
  | 'DATABASE_ID_ASC'
  | 'DATABASE_ID_DESC'
  | 'API_ID_ASC'
  | 'API_ID_DESC';
export type ApiModulesOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'DATABASE_ID_ASC'
  | 'DATABASE_ID_DESC'
  | 'API_ID_ASC'
  | 'API_ID_DESC'
  | 'NAME_ASC'
  | 'NAME_DESC'
  | 'DATA_ASC'
  | 'DATA_DESC';
export type DomainsOrderBy =
  | 'PRIMARY_KEY_ASC'
  | 'PRIMARY_KEY_DESC'
  | 'NATURAL'
  | 'ID_ASC'
  | 'ID_DESC'
  | 'DATABASE_ID_ASC'
  | 'DATABASE_ID_DESC'
  | 'API_ID_ASC'
  | 'API_ID_DESC'
  | 'SITE_ID_ASC'
  | 'SITE_ID_DESC'
  | 'SUBDOMAIN_ASC'
  | 'SUBDOMAIN_DESC'
  | 'DOMAIN_ASC'
  | 'DOMAIN_DESC';